refactor: consolidate receipt utils
